refactor(ghost): refactor decoration and editor config
The `EditorConfiguration` class has been refactored into a standalone function `getEditorConfiguration`.
The `SVGDecorationBuilder` class has been removed and its functionality consolidated into a new utility function `createSVGDecorationType`.
These changes simplify the structure and improve reusability of these utilities.

BREAKING CHANGE: The `EditorConfiguration` class is no longer available; use the `getEditorConfiguration` function instead. The `SVGDecorationBuilder` class has been removed; use the `createSVGDecorationType` function for creating SVG decorations.
