+++ .changeset/metal-corners-scream.md
@@ -1,0 +1,5 @@
+---
+"kilo-code": minor
+---
+
+Inline assist / autocomplete suggestions now support colorized code highlighting
